Refactored WebSocket-related tests
Extract WebSocket-related test infrastructure from PippoExtension to PippoWebSocketExtension
and make websocket-jetty-client dependency optional.
